docker_lambda/requirements.txt
psycopg2-binary==2.9.4
boto3==1.24.93
pandas